requirements.txt
asgiref==3.2.9
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
Django==3.0.7
get==2019.4.13
post==2019.4.13
public==2019.4.13
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
soupsieve==2.0.1
sqlparse==0.3.1
wincertstore==0.2
